@@ -9,42 +15,40 @@ classifiers =
     Operating System :: OS Independent
     Natural Language :: Polish
     Programming Language :: Python
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.4
     Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
     Programming Language :: Python :: Implementation :: CPython
-    Programming Language :: Python :: Implementation :: PyPy
     Programming Language :: Python :: Implementation :: Stackless
+    Programming Language :: Python :: Implementation :: PyPy
     Programming Language :: Python :: Implementation :: IronPython
     Programming Language :: Python :: Implementation :: Jython
     Topic :: Software Development :: Libraries :: Python Modules
 
 [options]
-python_requires = >=2.7.0,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+python_requires = >=3.5.0
 setup_requires =
-    setuptools>=39.2.0
+    setuptools>=41.2.0
 install_requires =
-    setuptools>=39.2.0
-    annotate
+    setuptools>=41.2.0
     libpcap
 packages = find:
+package_dir =
+    = src
+    # pcap.tests = tests
 zip_safe = True
 
 [options.packages.find]
-include = pcap, pcap.*
+where = src
 
 [options.extras_require]
 doc =
-    Sphinx>=1.6.7
+    Sphinx>=2.2.1
+    restructuredtext-lint>=1.3.0
 test =
-    unittest2;python_version<'3.0'
-    mock>=2.0.0;python_version<'3.0'
-    win_inet_pton>=1.0.1;python_version<'3.0'
-    dpkt>=1.9.1
+    dpkt>=1.9.2
 
 [sdist]
 formats = zip
@@ -54,7 +58,119 @@ universal = True
 
 [build_sphinx]
 source-dir = docs
-build-dir = docs/_build
+build-dir = build/docs
 builder = html
-all_files = True
+all-files = True
 warning-is-error = True
+
+#
+# Configuration(s) for tox
+#
+
+[tox:tox]
+envlist = py{35,36,37,38}, coverage, docs
+minversion = 3.13.2
+requires =
+    tox-venv>=0.4.0
+    {[base]setup_requires}
+skip_missing_interpreters = true
+
+[base]
+setup_requires =
+    pip>=19.3.1
+    setuptools>=41.6.0
+    wheel>=0.33.6
+packagesubdir = pcap
+
+[testenv]
+commands =
+    {envpython} --version
+    {envpython} -B -m tests {posargs}
+extras = test
+deps =
+    {[base]setup_requires}
+
+[testenv:prepare]
+basepython = python3.7
+skip_install = true
+whitelist_externals =
+    cmd
+    .build
+commands =
+    cmd /C if exist .build.cmd .build
+
+[testenv:coverage]
+basepython = python3.7
+commands =
+    {envpython} -m coverage erase
+    -{envpython} -B -m coverage run -m tests {posargs}
+    {envpython} -m coverage report
+    {envpython} -m coverage html
+extras = test
+deps =
+    {[testenv]deps}
+    coverage>=4.5.4
+
+[testenv:docs]
+basepython = python3.7
+commands =
+   #{envpython} -m sphinx.apidoc -f {envsitepackagesdir}/{[base]packagesubdir}
+    {envpython} setup.py -v build_sphinx -b html -E
+    {envpython} setup.py -v build_sphinx -b linkcheck
+    {envpython} setup.py -v build_sphinx -b doctest
+extras = doc
+
+[testenv:lint]
+basepython = python3.7
+commands =
+    {envpython} -m flake8 {envsitepackagesdir}/{[base]packagesubdir}/
+extras =
+deps =
+    {[testenv]deps}
+    flake8>=3.7.9
+    flake8-docstrings>=1.5.0
+    pep8-naming>=0.9.0
+    flake8-builtins>=1.4.1
+
+[coverage:run]
+source =
+    pcap
+    tests
+data_file = .tox/coverage/.coverage
+
+[coverage:report]
+exclude_lines =
+    # Regexes
+    # Have to re-enable the standard pragma
+    pragma: no cover
+    raise (AssertionError|NotImplementedError)
+    if\s+(0|False)\s*:
+    if\s+__name__.*\s*==\s*['"]__main__['"]\s*:
+    ^\s*@unittest\.skip\(
+    ^\s*@unittest\.skipIf\(sys\.platform\.startswith\("win"\)
+omit =
+    tests/run.py
+skip_covered = False
+
+[coverage:html]
+directory = .tox/coverage/.coverage.html
+
+[flake8]
+filename = *.py,*.pyx
+#include = tests
+#exclude = .tox,*.egg,.git,_build,.hypothesis
+max-line-length = 99
+ignore = E126,E203,E221,E251,E302,E701,E702,E731,
+         E122,E127,E128,E222,E272,E241,E266, E226,E704,
+         D100, D101, D102, D103, D104, D400,
+         N806, N802, N803, N801,
+         I100
+# (e.g. E4,W) default: E121,E123, 126, 226,E24, 704
+#select = 
+#select = E,W,F,N,I
+output-file = .tox/lint/flake8out.txt
+count = True
+#show-pep8,
+#show-source
+#verbose
+#quiet
