@@ -1,6 +1,6 @@
-# Copyright (c) 2016-2018, Adam Karpierz
+# Copyright (c) 2016-2019, Adam Karpierz
 # Licensed under the BSD license
-# http://opensource.org/licenses/BSD-3-Clause
+# https://opensource.org/licenses/BSD-3-Clause/
 
 """\
 packet capture library
@@ -12,8 +12,7 @@ accessible through this mechanism.
 
 # Based on pcap.pyx
 
-from __future__ import absolute_import
-
+from typing import Optional, Tuple, List
 import sys
 import struct
 import ctypes as ct
@@ -32,6 +31,7 @@ from libpcap import DLT_PFLOG
 from libpcap import DLT_RAW
 from libpcap import DLT_LOOP
 from libpcap import PCAP_D_INOUT, PCAP_D_IN, PCAP_D_OUT
+from libpcap import PCAP_TSTAMP_PRECISION_MICRO, PCAP_TSTAMP_PRECISION_NANO
 
 import libpcap as _pcap
 from . import _pcap_ex
@@ -52,26 +52,22 @@ dltoff = {
 }
 
 
-class bpf(object):
+class bpf:
 
     """bpf(filter, dlt=DLT_RAW) -> BPF filter object"""
 
     fcode = _pcap.bpf_program()
 
     def __init__(self, filter, dlt=DLT_RAW): # char *filter
-
         if _pcap_ex.compile_nopcap(65535, dlt, ct.byref(self.fcode),
                                    ct.c_char_p(filter), 1, 0) < 0:
             raise IOError("bad filter")
 
     def __del__(self):
-
         _pcap.freecode(ct.byref(self.fcode))
 
-    def filter(self, buf):
-
+    def filter(self, buf) -> bool:
         """Return boolean match for buf against our filter."""
-
         try:
             buf  = memoryview(buf).tobytes()
             size = len(buf)
@@ -82,9 +78,9 @@ class bpf(object):
                                 size, size) != 0
 
 
-class pcap(object):
+class pcap:
 
-    """pcap(name=None, snaplen=65535, promisc=True, timeout_ms=None, immediate=False)  -> packet capture object
+    """pcap(name=None, snaplen=65535, promisc=True, timeout_ms=None, immediate=False, timestamp_in_ns=False)  -> packet capture object
 
     Open a handle to a packet capture descriptor.
 
@@ -97,10 +93,12 @@ class pcap(object):
                   (in milliseconds) is reached and no packets were received
                   (Default: no timeout)
     immediate -- disable buffering, if possible
+    timestamp_in_ns -- report timestamps in integer nanoseconds
     """
 
     def __init__(self, name=None, snaplen=65535, promisc=True,
-                 timeout_ms=0, immediate=False):
+                 timeout_ms=0, immediate=False, rfmon=False,
+                 timestamp_in_ns=False):
 
         global dltoff
 
@@ -109,79 +107,134 @@ class pcap(object):
         if not name:
             cname = _pcap_ex.lookupdev(self.__ebuf)
             if cname is None:
-                raise OSError(str(self.__ebuf.value.decode("utf-8", "ignore")))
+                raise OSError(self.__ebuf.value.decode("utf-8", "ignore"))
         else:
             cname = name.encode("utf-8")
 
-        self.__pcap = _pcap.open_offline(cname, self.__ebuf)
+        try:
+            _pcap.open_offline_with_tstamp_precision
+        except AttributeError:
+            self.__pcap = _pcap.open_offline(cname, self.__ebuf)
+        else:
+            self.__pcap = _pcap.open_offline_with_tstamp_precision(cname,
+                                                                   PCAP_TSTAMP_PRECISION_NANO,
+                                                                   self.__ebuf)
         if not self.__pcap:
-            self.__pcap = _pcap.open_live(_pcap_ex.name(cname), snaplen, promisc,
-                                          timeout_ms, self.__ebuf)
+            self.__pcap = _pcap.create(_pcap_ex.name(cname), self.__ebuf)
+            def check_return(ret, descrip):
+                if ret != 0:
+                    raise OSError("{} failed to execute".format(descrip))
+            check_return(_pcap.set_snaplen(self.__pcap, snaplen), "Set snaplength")
+            check_return(_pcap.set_promisc(self.__pcap, promisc), "Set promiscuous mode")
+            check_return(_pcap.set_timeout(self.__pcap, timeout_ms), "Set timeout")
+            try:
+                _pcap.set_immediate_mode
+            except AttributeError as exc:
+                if immediate: raise exc
+            else:
+                check_return(_pcap.set_immediate_mode(self.__pcap, immediate),
+                                                      "Set immediate mode")
+            try:
+                _pcap.set_rfmon
+            except AttributeError as exc:
+               if rfmon: raise exc
+            else:
+                check_return(_pcap.set_rfmon(self.__pcap, rfmon), "Set monitor mode")
+            # Ask for nano-second precision, but don't fail if not available.
+            try:
+                _pcap.set_tstamp_precision
+            except AttributeError:
+                pass
+            else:
+                _pcap.set_tstamp_precision(self.__pcap, PCAP_TSTAMP_PRECISION_NANO)
+            if _pcap.activate(self.__pcap) != 0:
+                raise OSError("Activateing packet capture failed. "
+                              "Error returned by packet capture library "
+                              "was {}".format(self.geterr()))
         if not self.__pcap:
-            raise OSError(str(self.__ebuf.value.decode("utf-8", "ignore")))
+            raise OSError(self.__ebuf.value.decode("utf-8", "ignore"))
 
         self.__name   = cname
         self.__filter = b""
+        self.__timestamp_in_ns = timestamp_in_ns
+        try:
+            _pcap.get_tstamp_precision
+        except AttributeError:
+            precision = PCAP_TSTAMP_PRECISION_MICRO
+        else:
+            precision = _pcap.get_tstamp_precision(self.__pcap)
+        if precision == PCAP_TSTAMP_PRECISION_MICRO:
+            self.__precision_scale    = 1.0e-6
+            self.__precision_scale_ns = 1000
+        elif precision == PCAP_TSTAMP_PRECISION_NANO:
+            self.__precision_scale    = 1.0e-9
+            self.__precision_scale_ns = 1
+        else:
+            raise OSError("couldn't determine timestamp precision")
         try:
             self.__dloff = dltoff[_pcap.datalink(self.__pcap)]
         except KeyError:
-            self.__dloff = 0  # AK: added
+            self.__dloff = 0  # <AK>: added
 
         if immediate and _pcap_ex.immediate(self.__pcap) < 0:
             raise OSError("couldn't enable immediate mode")
 
     def __del__(self):
-
         try:
             if self.__pcap: _pcap.close(self.__pcap)
         except:
             pass
 
     @property
-    def name(self):
-
+    def name(self) -> str:
         """Network interface or dumpfile name."""
-
-        return str(self.__name.decode("utf-8"))
+        return self.__name.decode("utf-8")
 
     @property
     def snaplen(self):
-
         """Maximum number of bytes to capture for each packet."""
-
         return _pcap.snapshot(self.__pcap)
 
     @property
     def dloff(self):
-
         """Datalink offset (length of layer-2 frame header)."""
-
         return self.__dloff
 
     @property
-    def filter(self):
-
+    def filter(self) -> str:
         """Current packet capture filter."""
-
-        return str(self.__filter.decode("utf-8"))
+        return self.__filter.decode("utf-8")
 
     @property
     def fd(self):
-
         """File descriptor (or Win32 HANDLE) for capture handle."""
-
         return self.fileno()
 
+    @property
+    def precision(self):
+        """Precision of timestamps"""
+        try:
+            _pcap.get_tstamp_precision
+        except AttributeError:
+            return PCAP_TSTAMP_PRECISION_MICRO
+        return _pcap.get_tstamp_precision(self.__pcap)
+
+    @property
+    def timestamp_in_ns(self):
+        """Whether timestamps are returned in nanosecond units"""
+        return self.__timestamp_in_ns
+
     def fileno(self):
-
         """Return file descriptor (or Win32 HANDLE) for capture handle."""
-
         return _pcap_ex.fileno(self.__pcap)
 
+    def close(self):
+        """Explicitly close the underlying pcap handle"""
+        if self.__pcap: _pcap.close(self.__pcap)  # <AK>: condition added
+        self.__pcap = None
+
     def setfilter(self, value, optimize=1):
-
         """Set BPF-format packet capture filter."""
-
         fcode = _pcap.bpf_program()
         self.__filter = value.encode("utf-8")
         if _pcap.compile(self.__pcap, ct.byref(fcode), self.__filter, optimize, 0) < 0:
@@ -190,47 +243,47 @@ class pcap(object):
             raise OSError(self.geterr())
         _pcap.freecode(ct.byref(fcode))
 
-    def setdirection(self, direction):
-
+    def setdirection(self, direction) -> bool:
         """Set capture direction."""
-
-        return _pcap_ex.setdirection(self.__pcap, direction) == 0
+        try:
+            _pcap.setdirection
+        except AttributeError:
+            return False
+        return _pcap.setdirection(self.__pcap, direction) == 0
 
     def setnonblock(self, nonblock=True):
-
         """Set non-blocking capture mode."""
+        try:
+            _pcap.setnonblock
+        except AttributeError:
+            return
+        _pcap.setnonblock(self.__pcap, nonblock, self.__ebuf)
 
-        _pcap_ex.setnonblock(self.__pcap, nonblock, self.__ebuf)
-
-    def getnonblock(self):
-
+    def getnonblock(self) -> bool:
         """Return non-blocking capture mode as boolean."""
-
-        ret = _pcap_ex.getnonblock(self.__pcap, self.__ebuf)
+        try:
+            _pcap.getnonblock
+        except AttributeError:
+            return False
+        ret = _pcap.getnonblock(self.__pcap, self.__ebuf)
         if ret < 0:
-            raise OSError(str(self.__ebuf.value.decode("utf-8", "ignore")))
+            raise OSError(self.__ebuf.value.decode("utf-8", "ignore"))
         return ret != 0
 
     def datalink(self):
-
         """Return datalink type (DLT_* values)."""
-
         return _pcap.datalink(self.__pcap)
 
     def readpkts(self):
-
         """Return a list of (timestamp, packet) tuples received in one buffer."""
-
         pkts = []
         self.dispatch(-1, self.__add_pkts, pkts)
         return pkts
 
     def __add_pkts(self, ts, pkt, pkts):
-
         pkts.append((ts, pkt))
 
     def dispatch(self, cnt, callback, *args):
-
         """Collect and process packets with a user callback,
         return the number of packets processed, or 0 for a savefile.
 
@@ -243,41 +296,35 @@ class pcap(object):
         callback -- function with (timestamp, pkt, *args) prototype
         *args    -- optional arguments passed to callback on execution
         """
-
         ctx = _pcap_handler_ctx()
-        ctx.callback = callback
-        ctx.args     = args
-        ctx.exc      = None
+        ctx.scale           = self.__precision_scale
+        ctx.scale_ns        = self.__precision_scale_ns
+        ctx.timestamp_in_ns = self.__timestamp_in_ns
+        ctx.callback        = callback
+        ctx.args            = args
+        ctx.exc             = None
         n = _pcap.dispatch(self.__pcap, cnt, _pcap_handler,
                            ct.cast(ct.pointer(ctx), ct.POINTER(ct.c_ubyte)))
         exc = ctx.exc
         if exc is not None:
-            if sys.version_info[0] < 3:
-                raise exc[0](exc[1])
-            else:
-                raise exc[0](exc[1]).with_traceback(exc[2])
+            raise exc[0](exc[1]).with_traceback(exc[2])
         return n
 
-    def sendpacket(self, buf):
-
+    def sendpacket(self, buf) -> int:
         """Send a raw network packet on the interface."""
-
         if _pcap.sendpacket(self.__pcap, buf, len(buf)) == -1:
             raise OSError(self.geterr())
         return len(buf)
 
-    def stats(self):
-
+    def stats(self) -> Tuple:
         """Return a 3-tuple of the total number of packets received,
         dropped, and dropped by the interface."""
-
         pstat = _pcap.stat()
         if _pcap.stats(self.__pcap, ct.byref(pstat)) < 0:
             raise OSError(self.geterr())
         return (pstat.ps_recv, pstat.ps_drop, pstat.ps_ifdrop)
 
     def loop(self, cnt, callback, *args):
-
         """Processing packets with a user callback during a loop.
         The loop can be exited when cnt value is reached
         or with an exception, including KeyboardInterrupt.
@@ -290,142 +337,153 @@ class pcap(object):
         callback -- function with (timestamp, pkt, *args) prototype
         *args    -- optional arguments passed to callback on execution
         """
-
+        scale    = self.__precision_scale
+        scale_ns = self.__precision_scale_ns
+        timestamp_in_ns = self.__timestamp_in_ns
         _pcap_ex.setup(self.__pcap)
-
+        hdr  = _pcap.pkthdr()
+        phdr = ct.pointer(hdr)
+        pkt  = ct.POINTER(ct.c_ubyte)()
         i = 1
         while True:
             # with nogil:
-            n, hdr, pkt = _pcap_ex.next(self.__pcap)
+            n = _pcap_ex.next_ex(self.__pcap, ct.byref(phdr), ct.byref(pkt))
             if n == 0:  # timeout
                 continue
             elif n == 1:
-                header = hdr[0]
-                callback(header.ts.tv_sec + (header.ts.tv_usec / 1000000.0),
-                         ct.cast(pkt, ct.POINTER(ct.c_char * header.caplen))[0].raw,
-                         *args)
+                hdr = phdr[0]
+                if timestamp_in_ns:
+                    callback((hdr.ts.tv_sec * 1000000000) + (hdr.ts.tv_usec * scale_ns),
+                             ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw,
+                             *args)
+                else:
+                    callback(hdr.ts.tv_sec + (hdr.ts.tv_usec * scale),
+                             ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw,
+                             *args)
             elif n == -1:
                 raise KeyboardInterrupt()
             elif n == -2:
                 break
-            #else:  # AK: added
+            #else:  # <AK>: added
             #   ??? what about other/unknown codes?
             if i == cnt: break
             i += 1
 
-    def geterr(self):
-
+    def geterr(self) -> Optional[str]:
         """Return the last error message associated with this handle."""
-
         errmsg = _pcap.geterr(self.__pcap)
-        return str(errmsg.decode("utf-8", "ignore")) if errmsg is not None else None
+        return errmsg.decode("utf-8", "ignore") if errmsg is not None else None
 
     def __iter__(self):
-
         _pcap_ex.setup(self.__pcap)
         return self
 
     def __next__(self):
-
+        scale    = self.__precision_scale
+        scale_ns = self.__precision_scale_ns
+        timestamp_in_ns = self.__timestamp_in_ns
+        hdr  = _pcap.pkthdr()
+        phdr = ct.pointer(hdr)
+        pkt  = ct.POINTER(ct.c_ubyte)()
         while True:
             # with nogil:
-            n, hdr, pkt = _pcap_ex.next(self.__pcap)
+            n = _pcap_ex.next_ex(self.__pcap, ct.byref(phdr), ct.byref(pkt))
             if n == 0:  # timeout
                 continue
             elif n == 1:
-                header = hdr[0]
-                return (header.ts.tv_sec + (header.ts.tv_usec / 1000000.0),
-                        ct.cast(pkt, ct.POINTER(ct.c_char * header.caplen))[0].raw)
+                hdr = phdr[0]
+                if timestamp_in_ns:
+                    return ((hdr.ts.tv_sec * 1000000000) + (hdr.ts.tv_usec * scale_ns),
+                            ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw)
+                else:
+                    return (hdr.ts.tv_sec + (hdr.ts.tv_usec * scale),
+                            ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw)
             elif n == -1:
                 raise KeyboardInterrupt()
             elif n == -2:
                 raise StopIteration
-            #else:  # AK: added
+            #else:  # <AK>: added
             #   ??? what about other/unknown codes?
 
-    if sys.version_info[0] < 3:
-        next = __next__
 
-
-def ex_name(foo):
-
-    cname = foo.encode("utf-8")        # AK: added
+def ex_name(foo) -> str:
+    cname = foo.encode("utf-8")   # <AK>: added
     cname = _pcap_ex.name(cname)
-    return str(cname.decode("utf-8"))  # AK: added
+    return cname.decode("utf-8")  # <AK>: added
 
 
-def lookupdev():
-
+def lookupdev() -> str:
     """Return the name of a network device suitable for sniffing."""
-
     ebuf = ct.create_string_buffer(_pcap.PCAP_ERRBUF_SIZE)
     name = _pcap_ex.lookupdev(ebuf)
     if name is None:
-        raise OSError(str(ebuf.value.decode("utf-8", "ignore")))
-    return str(name.decode("utf-8"))
+        raise OSError(ebuf.value.decode("utf-8", "ignore"))
+    return name.decode("utf-8")
 
 
-def findalldevs():
-
+def findalldevs() -> List[str]:
     """Return a list of capture devices."""
-
     devs = ct.POINTER(_pcap.pcap_if_t)()
     ebuf = ct.create_string_buffer(_pcap.PCAP_ERRBUF_SIZE)
     status = _pcap.findalldevs(ct.byref(devs), ebuf)
     if status:
-        raise OSError(str(ebuf.value.decode("utf-8", "ignore")))
-    retval = []
+        raise OSError(ebuf.value.decode("utf-8", "ignore"))
     if not devs:
-        return retval
-    try:  # AK added
+        return []
+    retval = []
+    try:  # <AK> added
         dev = devs
         while dev:
             dev = dev[0]
-            retval.append(str(dev.name.decode("utf-8")))
+            retval.append(dev.name.decode("utf-8"))
             dev = dev.next
     finally:
         _pcap.freealldevs(devs)
     return retval
 
 
-def lookupnet(dev):
-
+def lookupnet(dev: str) -> Tuple:
     """
     Return the address and the netmask of a given device
     as network-byteorder integers.
     """
-
-    dev   = dev.encode("utf-8")  # AK: added
+    dev   = dev.encode("utf-8")  # <AK>: added
     netp  = ct.c_uint()
     maskp = ct.c_uint()
     ebuf  = ct.create_string_buffer(_pcap.PCAP_ERRBUF_SIZE)
     status = _pcap.lookupnet(dev, ct.byref(netp), ct.byref(maskp), ebuf)
     if status:
-        raise OSError(str(ebuf.value.decode("utf-8", "ignore")))
-    return struct.pack("I", netp.value), struct.pack("I", maskp.value)
+        raise OSError(ebuf.value.decode("utf-8", "ignore"))
+    return (struct.pack("I", netp.value), struct.pack("I", maskp.value))
 
 
 @_pcap.pcap_handler
-def _pcap_handler(arg, hdr, pkt): # with gil:
-
+def _pcap_handler(arg, hdr, pkt):  # with gil:
     ctx = ct.cast(arg, ct.POINTER(_pcap_handler_ctx))[0]
     try:
-        header   = hdr[0]
+        hdr = hdr[0]
+        scale    = ctx.scale
+        scale_ns = ctx.scale_ns
         callback = ctx.callback
         args     = ctx.args
-        callback(header.ts.tv_sec + (header.ts.tv_usec / 1000000.0),
-                 ct.cast(pkt, ct.POINTER(ct.c_char * header.caplen))[0].raw,
-                 *args)
+        if ctx.timestamp_in_ns:
+            callback((hdr.ts.tv_sec * 1000000000) + (hdr.ts.tv_usec * scale_ns),
+                     ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw,
+                     *args)
+        else:
+            callback(hdr.ts.tv_sec + (hdr.ts.tv_usec * scale),
+                     ct.cast(pkt, ct.POINTER(ct.c_char * hdr.caplen))[0].raw,
+                     *args)
     except:
         ctx.exc = sys.exc_info()
 
 
 class _pcap_handler_ctx(ct.Structure):
     _fields_ = [
-    ("callback", ct.py_object),
-    ("args",     ct.py_object),
-    ("exc",      ct.py_object),
+    ("scale",           ct.c_double),
+    ("scale_ns",        ct.c_longlong),
+    ("timestamp_in_ns", ct.c_bool),
+    ("callback",        ct.py_object),
+    ("args",            ct.py_object),
+    ("exc",             ct.py_object),
 ]
-
-
-# eof
