@@ -1,24 +1,21 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
+# Configuration file for the Sphinx documentation builder.
 #
-# Documentation build configuration file, created by sphinx-quickstart.
-#
-# This file is execfile()d with the current directory set to its
-# containing dir.
-#
-# Note that not all possible configuration values are present in this
-# autogenerated file.
-#
-# All configuration values have a default; values that are commented out
-# serve to show the default.
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
 
 import sys
 import os
 
 top_dir = os.path.join(os.path.dirname(__file__), os.pardir)
 
-with open(os.path.join(top_dir, "pcap", "__about__.py")) as f:
+with open(os.path.join(top_dir, "src", "pcap", "__about__.py")) as f:
     class about: exec(f.read(), None)
+
+def setup(app):
+    pass
+
+# -- Path setup --------------------------------------------------------------
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -29,11 +26,21 @@ with open(os.path.join(top_dir, "pcap", "__about__.py")) as f:
 # sys.path.insert(0, os.path.abspath('.'))
 
 
-# -- General configuration ------------------------------------------------
+# -- Project information -----------------------------------------------------
+
+project   = about.__title__
+copyright = about.__copyright__
+author    = about.__author__
+
+# The full version, including alpha/beta/rc tags
+release = about.__version__
+
+
+# -- General configuration ---------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-# needs_sphinx = '1.0'
+needs_sphinx = '2.0.1'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -42,18 +49,18 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.doctest',
+    'sphinx.ext.intersphinx',
+   #'sphinx.ext.todo',
+   #'sphinx.ext.coverage',
     'sphinx.ext.ifconfig',
     'sphinx.ext.napoleon',
 ]
 
+# Needed for e.g. linkcheck builder
+tls_verify = False
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
-
-# The suffix(es) of source filenames.
-# You can specify multiple suffix as a list of string:
-#
-# source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
 
 # The encoding of source files.
 #
@@ -62,52 +69,21 @@ source_encoding = 'utf-8'
 # The master toctree document.
 master_doc = 'index'
 
-# General information about the project.
-project   = about.__title__
-project_  = project.replace('.', '_')
-copyright = about.__copyright__
-author    = about.__author__
-
-# The version info for the project you're documenting, acts as replacement for
-# |version| and |release|, also used in various other places throughout the
-# built documents.
-#
-# The short X.Y version.
-version = '{0.major}.{0.minor}'.format(about.__version_info__)
-# The full version, including alpha/beta/rc tags.
-release = about.__version__
-
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-#
-# This is also used if you do content translation via gettext catalogs.
-# Usually you set "language" from the command line for these cases.
-language = None
-
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-# This patterns also effect to html_static_path and html_extra_path
+# This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = 'sphinx'
 
-# If true, `todo` and `todoList` produce output, else they produce nothing.
-todo_include_todos = False
 
-
-# -- Options for HTML output ----------------------------------------------
+# -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
 html_theme = 'sphinxdoc'  # 'alabaster'
-
-# Theme options are theme-specific and customize the look and feel of a theme
-# further.  For a list of options available for each theme, see the
-# documentation.
-#
-# html_theme_options = {}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -115,79 +91,25 @@ html_theme = 'sphinxdoc'  # 'alabaster'
 html_static_path = ['_static']
 
 
-# -- Options for HTMLHelp output ------------------------------------------
+# -- Extension configuration -------------------------------------------------
 
-# Output file base name for HTML help builder.
-htmlhelp_basename = project_ + '_' + 'doc'
+# -- Options for autodoc extension -------------------------------------------
 
+autoclass_content = 'both'
+autodoc_member_order = 'bysource'
 
-# -- Options for LaTeX output ---------------------------------------------
+# -- Options for apidoc extension --------------------------------------------
 
-latex_elements = {
-    # The paper size ('letterpaper' or 'a4paper').
-    #
-    # 'papersize': 'letterpaper',
+apidoc_separate_modules = True
+apidoc_module_first = True
+apidoc_output_dir = 'api'
 
-    # The font size ('10pt', '11pt' or '12pt').
-    #
-    # 'pointsize': '10pt',
+# -- Options for intersphinx extension ---------------------------------------
 
-    # Additional stuff for the LaTeX preamble.
-    #
-    # 'preamble': '',
+# Example configuration for intersphinx: refer to the Python standard library.
+# intersphinx_mapping = {'https://docs.python.org/': None}
 
-    # Latex figure (float) alignment
-    #
-    # 'figure_align': 'htbp',
-}
+# -- Options for todo extension ----------------------------------------------
 
-# Grouping the document tree into LaTeX files. List of tuples
-# (source start file, target name, title,
-#  author, documentclass [howto, manual, or own class]).
-latex_documents = [
-    (master_doc, project_ + '.tex', project + ' Documentation',
-     author, 'manual'),
-]
-
-
-# -- Options for manual page output ---------------------------------------
-
-# One entry per manual page. List of tuples
-# (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, project_.lower(), project + ' Documentation',
-     [author], 1)
-]
-
-
-# -- Options for Texinfo output -------------------------------------------
-
-# Grouping the document tree into Texinfo files. List of tuples
-# (source start file, target name, title, author,
-#  dir menu entry, description, category)
-texinfo_documents = [
-    (master_doc, project_, project + ' Documentation',
-     author, project_, about.__summary__,
-     'Miscellaneous'),
-]
-
-
-# -- Options for Epub output ----------------------------------------------
-
-# Bibliographic Dublin Core info.
-epub_title = project
-epub_author = author
-epub_publisher = author
-epub_copyright = copyright
-
-# The unique identifier of the text. This can be a ISBN number
-# or the project homepage.
-#
-# epub_identifier = ''
-
-# A unique identification for the text.
-#
-# epub_uid = ''
-
-# A list of files that should not be packed into the epub file.
-epub_exclude_files = ['search.html']
+# If true, `todo` and `todoList` produce output, else they produce nothing.
+todo_include_todos = False
