@@ -1,6 +1,6 @@
-# Copyright (c) 2016-2018, Adam Karpierz
+# Copyright (c) 2016-2019, Adam Karpierz
 # Licensed under the BSD license
-# http://opensource.org/licenses/BSD-3-Clause
+# https://opensource.org/licenses/BSD-3-Clause/
 
 from .__about__ import * ; del __about__
 __url__ = __uri__
