@@ -1,9 +1,8 @@
-# Copyright (c) 2016-2018, Adam Karpierz
+# Copyright (c) 2016-2019, Adam Karpierz
 # Licensed under the BSD license
-# http://opensource.org/licenses/BSD-3-Clause
+# https://opensource.org/licenses/BSD-3-Clause/
 
-from __future__ import absolute_import
-
+from typing import Optional, Tuple
 import os
 import re
 import tempfile
@@ -11,7 +10,6 @@ import socket
 import select
 import ctypes as ct
 
-from annotate import annotate
 from libpcap._platform import is_windows, is_osx, defined
 from libpcap._platform import sockaddr_in
 import libpcap as _pcap
@@ -20,9 +18,7 @@ if not is_windows:
     libc = ct.cdll.LoadLibrary("/lib64/libc.so.6")
 
 
-@annotate(int, pcap=ct.POINTER(_pcap.pcap_t))
-def immediate(pcap):
-
+def immediate(pcap: ct.POINTER(_pcap.pcap_t)) -> int:
     if is_windows:
         return _pcap.setmintocopy(pcap, 1)
     elif defined("BIOCIMMEDIATE"):
@@ -40,9 +36,7 @@ def immediate(pcap):
         return 0
 
 
-@annotate(bytes, bytes)
-def name(name):
-
+def name(name: bytes) -> bytes:
     if is_windows:
 
         # XXX - translate from libdnet logical interface name to
@@ -62,7 +56,7 @@ def name(name):
         if ret == -1:
             return name
 
-        try:  # AK: added
+        try:  # <AK>: added
             i   = 0
             dev = devs
             while dev:
@@ -75,24 +69,20 @@ def name(name):
             _pcap.freealldevs(devs)
 
         return name
-
     else:
         return name
 
 
-@annotate((bytes, None), ebuf=ct.c_char_p)
-def lookupdev(ebuf):
-
+def lookupdev(ebuf: ct.c_char_p) -> Optional[bytes]:
     if is_windows:
 
         # Get all available devices.
-
         ret, devs = _findalldevs(ebuf)
         if ret == -1:
             return None
 
         name = None
-        try:  # AK added
+        try:  # <AK> added
             # Get first not 0.0.0.0 or 127.0.0.1 device
             dev = devs
             while dev:
@@ -119,9 +109,7 @@ def lookupdev(ebuf):
         return _pcap.lookupdev(ebuf)
 
 
-@annotate(int, pcap=ct.POINTER(_pcap.pcap_t))
-def fileno(pcap):
-
+def fileno(pcap: ct.POINTER(_pcap.pcap_t)) -> int:
     if is_windows:
         # XXX - how to handle savefiles?
         return _pcap.getevent(pcap)
@@ -133,60 +121,37 @@ def fileno(pcap):
             return _pcap.fileno(pcap)
 
 
-@annotate(pcap=ct.POINTER(_pcap.pcap_t))
-def setup(pcap):
-
+def setup(pcap: ct.POINTER(_pcap.pcap_t)):
     # XXX - hrr, this sux
     if is_windows:
         ct.windll.kernel32.SetConsoleCtrlHandler(__ctrl_handler, 1) #, TRUE)
     else:
-        #if 0
-        """
-        fd = _pcap.fileno(pcap)
-        n = fcntl(fd, F_GETFL, 0) | O_NONBLOCK  # ct.c_int
-        fcntl(fd, F_SETFL, n);
-        """
-        #endif
+        if 0:
+            import fcntl
+            fd = _pcap.fileno(pcap)
+            n = fcntl.fcntl(fd, F_GETFL, 0) | os.O_NONBLOCK
+            fcntl.fcntl(fd, F_SETFL, n)
         libc.signal(SIGINT, __signal_handler)
 
 
-@annotate(int, pcap=ct.POINTER(_pcap.pcap_t), direction=int)
-def setdirection(pcap, direction):
+"""/*
+1.2.3: PCAP_API int pcap_next_ex(pcap_t*, struct pcap_pkthdr**, const u_char**);
+def next_ex(pcap: ct.POINTER(_pcap.pcap_t), struct pcap_pkthdr** hdr, u_char** pkt) -> int:
 
-    try:
-        _pcap.setdirection
-    except AttributeError:
-        return -2
-    else:
-        return _pcap.setdirection(pcap, direction)
+        struct timeval tv = { 1, 0 };
+        int fd, n;
 
-
-@annotate(int, pcap=ct.POINTER(_pcap.pcap_t), ebuf=ct.c_char_p)
-def getnonblock(pcap, ebuf):
-
-    try:
-        _pcap.getnonblock
-    except AttributeError:
-        return 0
-    else:
-        return _pcap.getnonblock(pcap, ebuf)
-
-
-@annotate(pcap=ct.POINTER(_pcap.pcap_t), nonblock=int, ebuf=ct.c_char_p)
-def setnonblock(pcap, nonblock, ebuf):
-
-    try:
-        _pcap.setnonblock
-    except AttributeError:
-        pass
-    else:
-        _pcap.setnonblock(pcap, nonblock, ebuf)
-
-
-__hdr = _pcap.pkthdr()
-
-@annotate(int, pcap=ct.POINTER(_pcap.pcap_t))
-def next(pcap):
+            fd_set rfds;
+            FD_ZERO(&rfds);
+            FD_SET(fd, &rfds);
+            n = select(fd + 1, &rfds, NULL, NULL, &tv);
+            if n <= 0:
+                return n
+*/
+"""
+def next_ex(pcap: ct.POINTER(_pcap.pcap_t),
+            hdr: ct.POINTER(ct.POINTER(_pcap.pkthdr)),
+            pkt: ct.POINTER(ct.POINTER(ct.c_ubyte))) -> int:
 
     # return codes: 1 = pkt, 0 = timeout, -1 = error, -2 = EOF
 
@@ -194,63 +159,48 @@ def next(pcap):
 
     if is_windows:
 
-        hdr = ct.POINTER(_pcap.pkthdr)()
-        pkt = ct.POINTER(ct.c_ubyte)()
-
         if __got_signal:
             __got_signal = False
-            return -1, hdr, pkt
+            return -1  # error
 
-        ret = _pcap.next_ex(pcap, ct.byref(hdr), ct.byref(pkt))
-        return ret, hdr, pkt
+        return _pcap.next_ex(pcap, hdr, pkt)
 
     else:
-
-        global __hdr
 
         fd = _pcap.fileno(pcap)
 
         while True:
-
-            hdr = ct.POINTER(_pcap.pkthdr)()
-            pkt = ct.POINTER(ct.c_ubyte)()
-
             if __got_signal:
                 __got_signal = False
-                return -1, hdr, pkt
+                return -1  # error
 
-            pkt = ct.cast(_pcap.next(pcap, ct.byref(__hdr)), ct.POINTER(ct.c_ubyte))
-            if pkt:
+            pkt[0] = ct.cast(_pcap.next(pcap, hdr[0]), ct.POINTER(ct.c_ubyte))
+            if pkt[0]:
                 break
 
             if _pcap.file(pcap):
-                return -2, hdr, pkt
+                return -2  # EOF
 
             try:
                 rfds, wfds, efds = select.select([fd], [], [], 1.0)
             except:
-                return -1, hdr, pkt
+                return -1  # error
             if not rfds and not wfds and not efds:
-                return 0, hdr, pkt  # timeout
+                return 0  # timeout
 
-        return 1, ct.pointer(__hdr), pkt
+        return 1  # pkt (OK)
 
 
-@annotate(int, snaplen=int, linktype=int,
-          prog=ct.POINTER(_pcap.bpf_program),
-          buffer=ct.c_char_p, optimize=int, mask=int)
-def compile_nopcap(snaplen, linktype, prog, buffer, optimize, mask):
-
+def compile_nopcap(snaplen: int, linktype: int, prog: ct.POINTER(_pcap.bpf_program),
+                   buffer: ct.c_char_p, optimize: int, mask: int) -> int:
     try:
         _pcap.compile_nopcap
     except AttributeError:
-        ret = -1
-
         try:
             f = tempfile.NamedTemporaryFile("wb", prefix=".pypcap", suffix=".pcap",
                                             delete=False)
         except:
-            return ret
+            return -1
 
         try:
             with f:
@@ -267,15 +217,15 @@ def compile_nopcap(snaplen, linktype, prog, buffer, optimize, mask):
 
             ebuf = ct.create_string_buffer(_pcap.PCAP_ERRBUF_SIZE)
             pcap = _pcap.open_offline(f.name, ebuf)
-            if pcap:
-                try:
-                    ret = _pcap.compile(pcap, prog, buffer, optimize, mask)
-                finally:
-                    _pcap.close(pcap)
+            if not pcap:
+                return -1
+
+            try:
+                return _pcap.compile(pcap, prog, buffer, optimize, mask)
+            finally:
+                _pcap.close(pcap)
         finally:
             os.unlink(f.name)
-
-        return ret
     else:
         if defined("__NetBSD__"):
             # We love consistent interfaces
@@ -291,11 +241,9 @@ if is_windows:
 
     import ctypes.wintypes
 
-    @annotate(tuple((int, ct.POINTER(_pcap.pcap_if_t))), ebuf=ct.c_char_p)
-    def _findalldevs(ebuf):
+    def _findalldevs(ebuf: ct.c_char_p) -> Tuple[int, ct.POINTER(_pcap.pcap_if_t)]:
 
         # XXX - set device list in libdnet order.
-
         devs = ct.POINTER(_pcap.pcap_if_t)()
         ret = _pcap.findalldevs(ct.byref(devs), ebuf)
         if ret == -1:
@@ -314,18 +262,13 @@ if is_windows:
 
     @ct.WINFUNCTYPE(ct.wintypes.BOOL, ct.wintypes.DWORD)
     def __ctrl_handler(sig):
-
         global __got_signal
         __got_signal = True
-        return 1 # TRUE
+        return 1  # TRUE
 
 else:
 
     @ct.CFUNCTYPE(None, ct.c_int)
     def __signal_handler(sig):
-
         global __got_signal
         __got_signal = True
-
-
-# eof
