@@ -16,6 +16,12 @@
 #include <pcap.h>
 #include "pcap_ex.h"
 
+#ifndef HAVE_PCAP_TSTAMP_PRECISION
+# define PCAP_TSTAMP_PRECISION_MICRO 0
+# define PCAP_TSTAMP_PRECISION_NANO 1
+# define PCAP_ERROR_TSTAMP_PRECISION_NOTSUP -12
+#endif
+
 /* XXX - hack around older Python versions */
 #include "patchlevel.h"
 #if PY_VERSION_HEX < 0x02030000
@@ -46,7 +52,7 @@ _pcap_ex_findalldevs(pcap_if_t **dst, char *ebuf)
 {
 	pcap_if_t *pifs, *cur, *prev, *next;
 	int ret;
-	
+
 	if ((ret = pcap_findalldevs(&pifs, ebuf)) != -1) {
 		/* XXX - flip script like a dyslexic actor */
 		for (prev = NULL, cur = pifs; cur != NULL; ) {
@@ -166,7 +172,7 @@ pcap_ex_setup(pcap_t *pcap)
 #else
 #if 0
 	int fd, n;
-	
+
 	fd = pcap_fileno(pcap);
 	n = fcntl(fd, F_GETFL, 0) | O_NONBLOCK;
 	fcntl(fd, F_SETFL, n);
@@ -203,9 +209,41 @@ pcap_ex_getnonblock(pcap_t *pcap, char *ebuf)
 #endif
 }
 
+int
+pcap_ex_get_tstamp_precision(pcap_t *pcap)
+{
+#ifdef HAVE_PCAP_TSTAMP_PRECISION
+	return (pcap_get_tstamp_precision(pcap));
+#else
+	return (PCAP_TSTAMP_PRECISION_MICRO);
+#endif
+}
+
+int pcap_ex_set_tstamp_precision(pcap_t *pcap, int tstamp_precision)
+{
+#ifdef HAVE_PCAP_TSTAMP_PRECISION
+	return (pcap_set_tstamp_precision(pcap, tstamp_precision));
+#else
+	if (tstamp_precision == PCAP_TSTAMP_PRECISION_MICRO)
+		return (0);
+	else
+		return PCAP_ERROR_TSTAMP_PRECISION_NOTSUP;
+#endif
+}
+
+pcap_t *pcap_ex_open_offline_with_tstamp_precision(char *fname,
+          unsigned int precision, char *errbuf)
+{
+#ifdef HAVE_PCAP_TSTAMP_PRECISION
+	return (pcap_open_offline_with_tstamp_precision(fname, precision, errbuf));
+#else
+	return (pcap_open_offline(fname, errbuf));
+#endif
+}
+
 /* return codes: 1 = pkt, 0 = timeout, -1 = error, -2 = EOF */
 int
-pcap_ex_next(pcap_t *pcap, struct pcap_pkthdr **hdr, u_char **pkt)
+pcap_ex_next(pcap_t *pcap, struct pcap_pkthdr *hdr, u_char **pkt)
 {
 #ifdef _WIN32
 	if (__pcap_ex_gotsig) {
@@ -214,8 +252,6 @@ pcap_ex_next(pcap_t *pcap, struct pcap_pkthdr **hdr, u_char **pkt)
 	}
 	return (pcap_next_ex(pcap, hdr, pkt));
 #else
-	static u_char *__pkt;
-	static struct pcap_pkthdr __hdr;
 	struct timeval tv = { 1, 0 };
 	fd_set rfds;
 	int fd, n;
@@ -226,7 +262,7 @@ pcap_ex_next(pcap_t *pcap, struct pcap_pkthdr **hdr, u_char **pkt)
 			__pcap_ex_gotsig = 0;
 			return (-1);
 		}
-		if ((__pkt = (u_char *)pcap_next(pcap, &__hdr)) == NULL) {
+		if ((*pkt = (u_char *)pcap_next(pcap, hdr)) == NULL) {
 			if (pcap_file(pcap) != NULL)
 				return (-2);
 			FD_ZERO(&rfds);
@@ -237,9 +273,7 @@ pcap_ex_next(pcap_t *pcap, struct pcap_pkthdr **hdr, u_char **pkt)
 		} else
 			break;
 	}
-	*pkt = __pkt;
-	*hdr = &__hdr;
-	
+
 	return (1);
 #endif
 }
@@ -264,7 +298,7 @@ pcap_ex_compile_nopcap(int snaplen, int dlt, struct bpf_program *fp, char *str,
 	char path[] = "/tmp/.pypcapXXXXXX.pcap";
 	char ebuf[PCAP_ERRBUF_SIZE];
 	int ret = -1;
-	
+
 	mktemp(path);
 	if ((f = fopen(path, "w")) != NULL) {
 		hdr.magic = 0xa1b2c3d4;
@@ -276,7 +310,7 @@ pcap_ex_compile_nopcap(int snaplen, int dlt, struct bpf_program *fp, char *str,
 		hdr.linktype = dlt;
 		fwrite(&hdr, sizeof(hdr), 1, f);
 		fclose(f);
-	
+
 		if ((pc = pcap_open_offline(path, ebuf)) != NULL) {
 			ret = pcap_compile(pc, fp, str, optimize, netmask);
 			pcap_close(pc);
