@@ -1,8 +1,6 @@
-# Copyright (c) 2016-2018, Adam Karpierz
+# Copyright (c) 2016-2019, Adam Karpierz
 # Licensed under the BSD license
-# http://opensource.org/licenses/BSD-3-Clause
-
-from __future__ import absolute_import, print_function
+# https://opensource.org/licenses/BSD-3-Clause/
 
 import unittest
 import sys
@@ -13,7 +11,6 @@ from . import test_dir
 
 
 def test_suite(names=None, omit=("test", "testsniff")):
-
     from . import __name__ as pkg_name
     from . import __path__ as pkg_path
     import unittest
@@ -26,16 +23,10 @@ def test_suite(names=None, omit=("test", "testsniff")):
     return tests
 
 
-def main():
-
-    print("Running testsuite", "\n", file=sys.stderr)
-
-    try:
-        tests = test_suite(sys.argv[1:] or None)
-        result = unittest.TextTestRunner(verbosity=2).run(tests)
-    finally:
-        pass
-
+def main(argv=sys.argv):
+    print("Running tests", "\n", file=sys.stderr)
+    tests = test_suite(argv[1:] or None)
+    result = unittest.TextTestRunner(verbosity=2).run(tests)
     sys.exit(0 if result.wasSuccessful() else 1)
 
 
